--- src/app/landingpage/page.tsx
@@ -1,7 +1,0 @@
-import LandingPage from "./landingpage";
-
-export default async function Home() {
-  return (
-    <LandingPage />
-  );
-}
